TASK-56521 Fix Collecting Activity Tag Analytics (#42)
Prior to this change, the activity analytics listener assumes that all tags added are about activities only. Thus, when adding tags to other entity types, some exceptions are raised due to the fact that the activity with the identifier of other Entity Types doesn't exist. This fix will ensure to add analytics only for tags added on activities.
