+++ analytics-listeners/src/main/resources/conf/portal/configuration.xml
@@ -14,9 +14,9 @@
   <external-component-plugins>
     <target-component>org.exoplatform.services.listener.ListenerService</target-component>
     <component-plugin>
-      <name>exo.tag.added</name>
+      <name>metadata.tag.added</name>
       <set-method>addListener</set-method>
-      <type>org.exoplatform.analytics.listener.social.AnalyticsTagsListener</type>
+      <type>org.exoplatform.analytics.listener.social.AnalyticsActivityTagsListener</type>
     </component-plugin>
   </external-component-plugins>
